% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_add.R
\name{mock_person_ids}
\alias{mock_person_ids}
\title{Participant ids in the mock database}
\usage{
mock_person_ids(con)
}
\arguments{
\item{con}{A connection from \code{\link[=mock_aou_connect]{mock_aou_connect()}}.}
}
\value{
An integer vector of \code{person_id}s.
}
\description{
Participant ids in the mock database
}
